add loginController and update security.yaml
